auth_service/auth_service/requirements.txt
Django==4.2.16
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.2
python-dotenv